+++ src/Gate.IO.Api/TradFi/Enums/GateTradFiKlineInterval.cs
@@ -12,10 +12,22 @@ public enum GateTradFiKlineInterval
     OneMinute = 1,
 
     /// <summary>
+    /// Represents the Five Minutes value.
+    /// </summary>
+    [Map("5m")]
+    FiveMinutes = 8,
+
+    /// <summary>
     /// Represents the Fifteen Minutes value.
     /// </summary>
     [Map("15m")]
     FifteenMinutes = 2,
+
+    /// <summary>
+    /// Represents the Thirty Minutes value.
+    /// </summary>
+    [Map("30m")]
+    ThirtyMinutes = 9,
 
     /// <summary>
     /// Represents the One Hour value.
@@ -46,4 +58,10 @@ public enum GateTradFiKlineInterval
     /// </summary>
     [Map("30d")]
     ThirtyDays = 7,
+
+    /// <summary>
+    /// Represents the One Month value.
+    /// </summary>
+    [Map("1M")]
+    OneMonth = 10,
 }