+++ src/Gate.IO.Api/TradFi/Enums/GateTradFiStreamPositionSide.cs
@@ -1,0 +1,17 @@
+namespace Gate.IO.Api.TradFi;
+
+/// <summary>
+/// TradFi stream position side.
+/// </summary>
+public enum GateTradFiStreamPositionSide
+{
+    /// <summary>
+    /// Short or sell side.
+    /// </summary>
+    Short = 1,
+
+    /// <summary>
+    /// Long or buy side.
+    /// </summary>
+    Long = 2,
+}